refactor(package): domain 패키지와 ui 패키지를 나눈다
